Creating separate component for cards
This should make them reusable.
