+++ src/index.css
@@ -561,51 +561,6 @@ body {
   justify-content: space-between;
   margin: 0 -20px; }
 
-.card {
-  background-color: #f7f7f7;
-  border: 1px solid #ddd;
-  border-radius: 3px;
-  box-shadow: 0 2px 4px #e6e6e6;
-  cursor: pointer;
-  flex-basis: 15em;
-  flex-grow: 1;
-  margin: 0 1em 1.5em 1em;
-  position: relative;
-  transition: all 0.2s ease-in-out; }
-  .card .card-image {
-    background-color: #66C3EF;
-    height: 200px;
-    max-height: 200px;
-    overflow: hidden; }
-    .card .card-image img {
-      border-top-left-radius: 3px;
-      border-top-right-radius: 3px;
-      opacity: 1;
-      transition: all 0.2s ease-in-out;
-      width: 100%; }
-  .card .card-header {
-    background-color: #f7f7f7;
-    border-bottom: 1px solid #ddd;
-    border-radius: 3px 3px 0 0;
-    font-weight: bold;
-    line-height: 1.5em;
-    padding: 0.5em 0.75em;
-    transition: all 0.2s ease-in-out; }
-  .card .card-copy {
-    font-size: 0.9em;
-    line-height: 1.5em;
-    padding: 0.75em 0.75em; }
-    .card .card-copy p {
-      margin: 0 0 0.75em; }
-  .card:focus, .card:hover {
-    cursor: pointer; }
-    .card:focus img, .card:hover img {
-      opacity: 0.7; }
-  .card:active {
-    background-color: #f7f7f7; }
-    .card:active .card-header {
-      background-color: #f7f7f7; }
-
 .nav_items {
   display: flex;
   flex-wrap: wrap;
